Move a few more shared files
