+++ static/app/views/dashboards/widgets/timeSeriesWidget/formatTooltipValue.tsx
