+++ static/app/views/dashboards/widgets/timeSeriesWidget/formatYAxisValue.spec.tsx
