+++ static/app/views/dashboards/widgets/timeSeriesWidget/formatYAxisValue.tsx
